refactor(presentation): 리소스 경로 presentation으로 이동
