+++ android/festago/app/src/main/java/com/festago/festago/FestagoApplication.kt
@@ -4,6 +4,7 @@ import android.app.Application
 import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.content.Context
+import com.festago.festago.presentation.R
 import com.festago.festago.presentation.fcm.FcmMessageType
 import com.kakao.sdk.common.KakaoSdk
 import dagger.hilt.android.HiltAndroidApp
