+++ android/festago/gradle.properties
@@ -22,3 +22,5 @@ kotlin.code.style=official
 # thereby reducing the size of the R class for that library
 android.nonTransitiveRClass=true
 android.defaults.buildfeatures.buildconfig=true
+android.enableJetifier=true
+android.databinding.enableV2=true
