+++ android/festago/presentation/src/main/java/com/festago/festago/presentation/ui/studentverification/StudentVerificationActivity.kt
@@ -5,8 +5,8 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import com.festago.festago.R
-import com.festago.festago.databinding.ActivityStudentVerificationBinding
+import com.festago.festago.presentation.R
+import com.festago.festago.presentation.databinding.ActivityStudentVerificationBinding
 import com.festago.festago.presentation.ui.customview.OkDialogFragment
 import com.festago.festago.presentation.ui.home.HomeActivity
 import com.festago.festago.presentation.util.repeatOnStarted
