+++ android/festago/presentation/src/main/java/com/festago/festago/presentation/ui/ticketentry/TicketEntryActivity.kt
@@ -8,7 +8,7 @@ import androidx.activity.OnBackPressedCallback
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.content.res.ResourcesCompat
-import com.festago.festago.databinding.ActivityTicketEntryBinding
+import com.festago.festago.presentation.databinding.ActivityTicketEntryBinding
 import com.festago.festago.presentation.fcm.TicketEntryService
 import com.festago.festago.presentation.util.repeatOnStarted
 import com.google.zxing.BarcodeFormat
