+++ android/festago/presentation/src/main/java/com/festago/festago/presentation/ui/tickethistory/TicketHistoryViewHolder.kt
@@ -3,7 +3,7 @@ package com.festago.festago.presentation.ui.tickethistory
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.festago.festago.databinding.ItemTicketHistoryBinding
+import com.festago.festago.presentation.databinding.ItemTicketHistoryBinding
 
 class TicketHistoryViewHolder(
     val binding: ItemTicketHistoryBinding,
