+++ android/festago/presentation/src/main/java/com/festago/festago/presentation/ui/ticketreserve/viewHolder/TicketReserveHeaderViewHolder.kt
@@ -5,8 +5,8 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.engine.DiskCacheStrategy
-import com.festago.festago.R
-import com.festago.festago.databinding.ItemTicketReserveHeaderBinding
+import com.festago.festago.presentation.R
+import com.festago.festago.presentation.databinding.ItemTicketReserveHeaderBinding
 import com.festago.festago.presentation.ui.ticketreserve.ReservationFestivalUiState
 import java.time.format.DateTimeFormatter
