+++ android/festago/presentation/src/main/java/com/festago/festago/presentation/ui/ticketreserve/viewHolder/TicketReserveViewHolder.kt
@@ -3,9 +3,9 @@ package com.festago.festago.presentation.ui.ticketreserve.viewHolder
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.festago.festago.R
-import com.festago.festago.databinding.ItemTicketReserveBinding
 import com.festago.festago.model.TicketType
+import com.festago.festago.presentation.R
+import com.festago.festago.presentation.databinding.ItemTicketReserveBinding
 import com.festago.festago.presentation.ui.ticketreserve.TicketReserveItemUiState
 import java.time.format.DateTimeFormatter
